app/requirements.txt
Flask
protobuf
python-dotenv
requests_oauthlib
google-cloud-datastore
google-cloud-firestore
google-cloud-storage
gunicorn
tenacity
google-cloud-aiplatform